+++ CHANGELOG.md
@@ -4,7 +4,12 @@ Note: in the following, FID = wasatch.FeatureInterfaceDevice
 
 - 2026-??-?? 2.4.7 (enlighten-4.2.8?)
     - more JSON fixes
-    - read Andor capabilities (including camera type)
+    - bad pixel correction
+        - promote correct_bad_pixels from FeatureIdentificationDevice to InterfaceDevice
+        - support bad_pixels in AndorDevice
+        - support bad_pixels in BLEDevice
+    - AndorDevice
+        - read capabilities (including camera type)
 - 2026-08-06 2.4.6 (enlighten-4.2.7)
     - updated markdown docs
     - JSON